important additions to mkd: hasmultikey, index
implementing index necessitated implementing class toggling
(OrderedDict <-> tuple) for the _keyorder attribute, which in turn
necessitated implementing some helper methods to hide this
implementation detail from the rest of the class.  (This is a hint
that maybe this functionality should really be hidden within a
subclass of OrderedDict and/or tuple.)

Also, added an extended docstring/comment showing the space savings of
MKDs relative to the equivalent dict with tuples as keys.