small but significant change to dimension.py
I got rid of the possibility to pass a tuple of specs to the Dimension.index method; it was causing problems with composite dimensions (like ('cell_line', 'repno') in the current incarnation of the icbp45 data).  To make up for this, as an expedient I wrote in a new Dimension.indexes, whose only reason to be is to take a tuple of indices as argument.